sherd

+++ src/Minimarket/Sherd/Dto/Product/GetProductDto.cs
@@ -1,4 +1,4 @@
 ﻿namespace Sheard.Dto.Product
 {
-    public record GetProductDto(string ProductName, decimal Price, Guid ProductId, Guid CategoryId, DateTime CreateDateTime, DateTime ModifiDateTime);
+    public record GetProductDto(string ProductName, int Price, Guid ProductId, Guid CategoryId, DateTime CreateDateTime, DateTime ModifiDateTime);
 }
