+++ src/Minimarket/Sherd/Dto/Product/UpdateProductDto.cs
@@ -1,4 +1,4 @@
 ﻿namespace Sheard.Dto.Product
 {
-    public record UpdateProductDto(string ProductName, Guid CategoryId, decimal Price);
+    public record UpdateProductDto(string ProductName, Guid CategoryId, int Price);
 }
